+++ src/util.h
@@ -15,9 +15,9 @@ inline double reldist(numeric x, numeric y, double eps)
 }
 
 
-double uniform_distance(int n, const numeric* x, const numeric* y);
+double uniform_distance(size_t n, const numeric* x, const numeric* y);
 
-double relative_distance(int n, const numeric* x, const numeric* y, double eps);
+double relative_distance(size_t n, const numeric* x, const numeric* y, double eps);
 
 
 void inverse_permutation(int n, const int* restrict perm, int* restrict inv_perm);
